--- app/controllers/application_controller.rb
@@ -1,13 +1,0 @@
-class ApplicationController < ActionController::API
-  include Response
-  include ExceptionHandler
-
-  before_action :authorize_request
-  attr_reader :current_user
-
-  private
-
-    def authorize_request
-      @current_user = Auth::AuthorizeRequestService.call(request.headers)[:user]
-    end
-end
