+++ app/controllers/v1/application_controller.rb
@@ -1,0 +1,15 @@
+module V1
+  class ApplicationController < ActionController::API
+    include Response
+    include ExceptionHandler
+
+    before_action :authorize_request
+    attr_reader :current_user
+
+    private
+
+      def authorize_request
+        @current_user = Auth::AuthorizeRequestService.call(request.headers)[:user]
+      end
+  end
+end
