+++ spec/controllers/application_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe ApplicationController, type: :controller do
+RSpec.describe V1::ApplicationController, type: :controller do
   let!(:user) { create(:user) }
 
   describe '#authorize_request' do
